+++ devcontrol/global/startup.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+# @file devcontrol/global/startup.sh
+# @brief devcontrol startup script and functions
+echo "docker-learning Kairops (c) 2019"
+echo